moved gemini settlement code into settlement folder
