+++ cmd/settlement/gemini.go
@@ -2,25 +2,18 @@ package settlement
 
 import (
 	"context"
-	"encoding/base64"
-	"encoding/hex"
 	"encoding/json"
 	"errors"
 	"io/ioutil"
 	"path/filepath"
 	"strings"
-	"time"
 
 	"github.com/brave-intl/bat-go/cmd"
 	"github.com/brave-intl/bat-go/settlement"
-	"github.com/brave-intl/bat-go/utils/altcurrency"
+	geminisettlement "github.com/brave-intl/bat-go/settlement/gemini"
 	"github.com/brave-intl/bat-go/utils/clients/gemini"
 	appctx "github.com/brave-intl/bat-go/utils/context"
-	"github.com/brave-intl/bat-go/utils/cryptography"
 	"github.com/brave-intl/bat-go/utils/logging"
-	"github.com/rs/zerolog"
-	uuid "github.com/satori/go.uuid"
-	"github.com/shopspring/decimal"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
 )
@@ -92,15 +85,24 @@ func CheckStatusGeminiSettlement(cmd *cobra.Command, args []string) error {
 	if out == "" {
 		out = strings.TrimSuffix(input, filepath.Ext(input)) + "-finished.json"
 	}
+	sig, err := cmd.Flags().GetInt("sig")
+	if err != nil {
+		return err
+	}
+	allTxsInput, err := cmd.Flags().GetString("all-txs-input")
+	if err != nil {
+		return err
+	}
 	return GeminiUploadSettlement(
 		cmd.Context(),
 		"checkstatus",
 		input,
-		viper.GetInt("sig"),
-		viper.GetString("all-txs-input"),
+		sig,
+		allTxsInput,
 		out,
 	)
 }
+
 func init() {
 	// add complete and transform subcommand
 	GeminiSettlementCmd.AddCommand(UploadGeminiSettlementCmd)
@@ -112,92 +114,51 @@ func init() {
 	// setup the flags
 
 	// input (required by all)
-	GeminiSettlementCmd.PersistentFlags().String("input", "",
+	GeminiSettlementCmd.Flags().String("input", "",
 		"the file or comma delimited list of files that should be utilized")
-	cmd.Must(viper.BindPFlag("input", GeminiSettlementCmd.PersistentFlags().Lookup("input")))
+	cmd.Must(viper.BindPFlag("input", GeminiSettlementCmd.Flags().Lookup("input")))
 	cmd.Must(viper.BindEnv("input", "INPUT"))
-	cmd.Must(GeminiSettlementCmd.MarkPersistentFlagRequired("input"))
+	cmd.Must(GeminiSettlementCmd.MarkFlagRequired("input"))
 
 	// out (required by all with default)
-	GeminiSettlementCmd.PersistentFlags().String("out", "./gemini-settlement",
+	GeminiSettlementCmd.Flags().String("out", "./gemini-settlement",
 		"the location of the file")
-	cmd.Must(viper.BindPFlag("out", GeminiSettlementCmd.PersistentFlags().Lookup("out")))
+	cmd.Must(viper.BindPFlag("out", GeminiSettlementCmd.Flags().Lookup("out")))
 	cmd.Must(viper.BindEnv("out", "OUT"))
 
 	// txnID (required by transform)
-	UploadGeminiSettlementCmd.PersistentFlags().String("input", "",
+	UploadGeminiSettlementCmd.Flags().String("input", "",
 		"the signed transactions file")
-	cmd.Must(viper.BindPFlag("input", UploadGeminiSettlementCmd.PersistentFlags().Lookup("input")))
-	cmd.Must(UploadGeminiSettlementCmd.MarkPersistentFlagRequired("input"))
+	cmd.Must(viper.BindPFlag("input", UploadGeminiSettlementCmd.Flags().Lookup("input")))
+	cmd.Must(UploadGeminiSettlementCmd.MarkFlagRequired("input"))
 
-	UploadGeminiSettlementCmd.PersistentFlags().String("all-txs-input", "",
+	UploadGeminiSettlementCmd.Flags().String("all-txs-input", "",
 		"the original transactions file")
-	cmd.Must(viper.BindPFlag("all-txs-input", UploadGeminiSettlementCmd.PersistentFlags().Lookup("all-txs-input")))
-	cmd.Must(UploadGeminiSettlementCmd.MarkPersistentFlagRequired("all-txs-input"))
+	cmd.Must(viper.BindPFlag("all-txs-input", UploadGeminiSettlementCmd.Flags().Lookup("all-txs-input")))
+	cmd.Must(UploadGeminiSettlementCmd.MarkFlagRequired("all-txs-input"))
 
-	UploadGeminiSettlementCmd.PersistentFlags().Int("sig", 0,
+	UploadGeminiSettlementCmd.Flags().Int("sig", 0,
 		"the original transactions file")
-	cmd.Must(viper.BindPFlag("sig", UploadGeminiSettlementCmd.PersistentFlags().Lookup("sig")))
+	cmd.Must(viper.BindPFlag("sig", UploadGeminiSettlementCmd.Flags().Lookup("sig")))
 
 	// CheckStatusGeminiSettlementCmd
-	CheckStatusGeminiSettlementCmd.PersistentFlags().String("all-txs-input", "",
+	CheckStatusGeminiSettlementCmd.Flags().String("all-txs-input", "",
 		"the original transactions file")
-	cmd.Must(viper.BindPFlag("all-txs-input", CheckStatusGeminiSettlementCmd.PersistentFlags().Lookup("all-txs-input")))
-	cmd.Must(CheckStatusGeminiSettlementCmd.MarkPersistentFlagRequired("all-txs-input"))
+	cmd.Must(viper.BindPFlag("all-txs-input", CheckStatusGeminiSettlementCmd.Flags().Lookup("all-txs-input")))
+	cmd.Must(CheckStatusGeminiSettlementCmd.MarkFlagRequired("all-txs-input"))
 
-	CheckStatusGeminiSettlementCmd.PersistentFlags().StringP("input", "i", "",
+	CheckStatusGeminiSettlementCmd.Flags().StringP("input", "i", "",
 		"the original transactions file")
-	cmd.Must(viper.BindPFlag("input", CheckStatusGeminiSettlementCmd.PersistentFlags().Lookup("input")))
-	cmd.Must(CheckStatusGeminiSettlementCmd.MarkPersistentFlagRequired("input"))
+	cmd.Must(viper.BindPFlag("input", CheckStatusGeminiSettlementCmd.Flags().Lookup("input")))
+	cmd.Must(CheckStatusGeminiSettlementCmd.MarkFlagRequired("input"))
 
 	CheckStatusGeminiSettlementCmd.Flags().String("out", "",
 		"the output file name")
 	cmd.Must(viper.BindPFlag("out", CheckStatusGeminiSettlementCmd.Flags().Lookup("out")))
 
-	CheckStatusGeminiSettlementCmd.PersistentFlags().Int("sig", 0,
+	CheckStatusGeminiSettlementCmd.Flags().Int("sig", 0,
 		"signature to choose (for bulk endpoint usage)")
-	cmd.Must(viper.BindPFlag("sig", CheckStatusGeminiSettlementCmd.PersistentFlags().Lookup("sig")))
-}
-
-func categorizeResponse(
-	originalTransactions map[string]settlement.Transaction,
-	payout *gemini.PayoutResult,
-) (settlement.Transaction, string) {
-	original := originalTransactions[payout.TxRef]
-	key := "failed"
-	if payout.Result == "Error" {
-		original.Note = *payout.Reason
-	} else {
-		status := *payout.Status
-		key = "unknown"
-		if *payout.Status == "Pending" {
-			key = "pending"
-		} else if status == "Completed" {
-			key = "complete"
-		}
-	}
-	original.Status = key
-	tmp := altcurrency.BAT
-	original.AltCurrency = &tmp
-	original.Currency = tmp.String()
-	original.ProviderID = payout.TxRef
-	return original, key
-}
-
-func geminiSiftThroughResponses(
-	originalTransactions map[string]settlement.Transaction,
-	response *[]gemini.PayoutResult,
-) map[string][]settlement.Transaction {
-	transactions := make(map[string][]settlement.Transaction)
-
-	for _, payout := range *response {
-		original, key := categorizeResponse(
-			originalTransactions,
-			&payout,
-		)
-		transactions[key] = append(transactions[key], original)
-	}
-	return transactions
+	cmd.Must(viper.BindPFlag("sig", CheckStatusGeminiSettlementCmd.Flags().Lookup("sig")))
 }
 
 // GeminiUploadSettlement marks the settlement file as complete
@@ -239,7 +200,14 @@ func GeminiUploadSettlement(ctx context.Context, action string, inPath string, s
 	// create a map of the request transactions
 	transactionsMap := geminiMapTransactionsToID(settlementTransactions)
 
-	submittedTransactions, submitErr := geminiIterateRequest(ctx, action, geminiClient, signatureSwitch, bulkPayoutFiles, transactionsMap)
+	submittedTransactions, submitErr := geminisettlement.IterateRequest(
+		ctx,
+		action,
+		geminiClient,
+		signatureSwitch,
+		bulkPayoutFiles,
+		transactionsMap,
+	)
 	// write file for upload to eyeshade
 	logger.Info().
 		Str("files", outPath).
@@ -256,198 +224,6 @@ func GeminiUploadSettlement(ctx context.Context, action string, inPath string, s
 		}
 	}
 	return submitErr
-}
-
-func geminiIterateRequest(
-	ctx context.Context,
-	action string,
-	geminiClient gemini.Client,
-	signatureSwitch int,
-	bulkPayoutFiles []string,
-	transactionsMap map[string]settlement.Transaction,
-) (*map[string][]settlement.Transaction, error) {
-
-	logger, err := appctx.GetLogger(ctx)
-	if err != nil {
-		_, logger = logging.SetupLogger(ctx)
-	}
-
-	submittedTransactions := make(map[string][]settlement.Transaction)
-
-	for _, bulkPayoutFile := range bulkPayoutFiles {
-		bytes, err := ioutil.ReadFile(bulkPayoutFile)
-		if err != nil {
-			logger.Error().Err(err).Msg("failed to read bulk payout file")
-			return &submittedTransactions, err
-		}
-
-		var geminiBulkPayoutRequestRequirements []gemini.PrivateRequestSequence
-		err = json.Unmarshal(bytes, &geminiBulkPayoutRequestRequirements)
-		if err != nil {
-			logger.Error().Err(err).Msg("failed unmarshal bulk payout file")
-			return &submittedTransactions, err
-		}
-
-		total := geminiComputeTotal(geminiBulkPayoutRequestRequirements)
-		for i, bulkPayoutRequestRequirements := range geminiBulkPayoutRequestRequirements {
-			blockProgress := geminiComputeTotal(geminiBulkPayoutRequestRequirements[:i+1])
-			if action == "upload" {
-				submittedTransactions, err = submitBulkPayoutTransactions(
-					ctx,
-					transactionsMap,
-					submittedTransactions,
-					bulkPayoutRequestRequirements,
-					geminiClient,
-					len(bulkPayoutFiles),
-					blockProgress,
-					signatureSwitch,
-				)
-				if err != nil {
-					logger.Error().Err(err).Msg("failed to submit bulk payout transactions")
-					return nil, err
-				}
-			} else if action == "checkstatus" {
-				submittedTransactions, err = checkPayoutTransactionsStatus(
-					ctx,
-					transactionsMap,
-					submittedTransactions,
-					bulkPayoutRequestRequirements,
-					geminiClient,
-					total,
-					blockProgress,
-				)
-				if err != nil {
-					logger.Error().Err(err).Msg("falied to check payout transactions status")
-					return nil, err
-				}
-			}
-		}
-	}
-	return &submittedTransactions, nil
-}
-
-func geminiComputeTotal(geminiBulkPayoutRequestRequirements []gemini.PrivateRequestSequence) int {
-	if len(geminiBulkPayoutRequestRequirements) == 0 {
-		return 0
-	}
-
-	firstLen := len(geminiBulkPayoutRequestRequirements[0].Base.Payouts)
-	blockLen := len(geminiBulkPayoutRequestRequirements)
-	lastLen := len(geminiBulkPayoutRequestRequirements[blockLen-1].Base.Payouts)
-	total := blockLen * firstLen
-	if blockLen > 1 {
-		total += lastLen
-	}
-	return total
-}
-
-func checkPayoutTransactionsStatus(
-	ctx context.Context,
-	transactionsMap map[string]settlement.Transaction,
-	submittedTransactions map[string][]settlement.Transaction,
-	bulkPayoutRequestRequirements gemini.PrivateRequestSequence,
-	geminiClient gemini.Client,
-	total int,
-	blockProgress int,
-) (map[string][]settlement.Transaction, error) {
-	logger, err := appctx.GetLogger(ctx)
-	if err != nil {
-		_, logger = logging.SetupLogger(ctx)
-	}
-
-	APIKey := bulkPayoutRequestRequirements.APIKey
-	base := bulkPayoutRequestRequirements.Base
-	clientID := base.OauthClientID
-	for j, payout := range base.Payouts {
-		result, err := geminiClient.CheckTxStatus(
-			ctx,
-			APIKey,
-			clientID,
-			payout.TxRef,
-		)
-		if err != nil {
-			return nil, err
-		}
-		original, key := categorizeResponse(
-			transactionsMap,
-			result,
-		)
-		submittedTransactions[key] = append(submittedTransactions[key], original)
-
-		prog := blockProgress + j
-		logging.SubmitProgress(ctx, prog, total)
-		logger.Debug().
-			Int("total", total).
-			Int("progress", prog).
-			Str("key", key).
-			Str("tx_ref", payout.TxRef).
-			Msg("parameters used")
-	}
-	return submittedTransactions, err
-}
-
-func submitBulkPayoutTransactions(
-	ctx context.Context,
-	transactionsMap map[string]settlement.Transaction,
-	submittedTransactions map[string][]settlement.Transaction,
-	bulkPayoutRequestRequirements gemini.PrivateRequestSequence,
-	geminiClient gemini.Client,
-	total int,
-	blockProgress int,
-	signatureSwitch int,
-) (map[string][]settlement.Transaction, error) {
-	logger, err := appctx.GetLogger(ctx)
-	if err != nil {
-		_, logger = logging.SetupLogger(ctx)
-	}
-	logging.SubmitProgress(ctx, blockProgress, total)
-	// make sure payload is parsable
-	// upload the bulk payout
-	sig := bulkPayoutRequestRequirements.Signatures[signatureSwitch]
-	decodedSig, err := hex.DecodeString(sig)
-	if err != nil {
-		logger.Error().Err(err).Msg("failed decode signature")
-		return submittedTransactions, err
-	}
-	base := bulkPayoutRequestRequirements.Base
-	presigner := cryptography.NewPresigner(decodedSig)
-	base.Nonce = base.Nonce + int64(signatureSwitch)
-
-	logger.Debug().
-		Int("total", total).
-		Int("progress", blockProgress).
-		Int64("nonce", base.Nonce).
-		Int("signature switch", signatureSwitch).
-		Msg("parameters used")
-
-	serialized, err := json.Marshal(base)
-	if err != nil {
-		return submittedTransactions, err
-	}
-	payload := base64.StdEncoding.EncodeToString(serialized)
-
-	logger.Debug().
-		Str("api key", bulkPayoutRequestRequirements.APIKey).
-		Str("signature", sig).
-		Msg("sending request")
-
-	response, err := geminiClient.UploadBulkPayout(
-		ctx,
-		bulkPayoutRequestRequirements.APIKey,
-		presigner,
-		payload,
-	)
-	<-time.After(time.Second)
-	if err != nil {
-		logger.Error().Err(err).Msg("error performing upload")
-		return submittedTransactions, err
-	}
-	// collect all successful transactions to send to eyeshade
-	submitted := geminiSiftThroughResponses(transactionsMap, response)
-	for key, txs := range submitted {
-		submittedTransactions[key] = append(submittedTransactions[key], txs...)
-	}
-	return submittedTransactions, nil
 }
 
 // geminiMapTransactionsToID creates a map of guid's to transactions
@@ -487,55 +263,4 @@ func GeminiWriteRequests(outPath string, metadata *[][]gemini.PayoutPayload) err
 		return err
 	}
 	return ioutil.WriteFile(outPath, data, 0600)
-}
-
-// GeminiConvertTransactionsToGeminiPayouts converts transactions from antifraud to "payouts" for gemini
-func GeminiConvertTransactionsToGeminiPayouts(transactions *[]settlement.Transaction, txID string) (*[]gemini.PayoutPayload, decimal.Decimal) {
-	payouts := make([]gemini.PayoutPayload, 0)
-	total := decimal.NewFromFloat(0)
-	for i, tx := range *transactions {
-		tx.SettlementID = txID
-		(*transactions)[i] = tx
-		payout := gemini.SettlementTransactionToPayoutPayload(&tx)
-		total = total.Add(payout.Amount)
-		payouts = append(payouts, payout)
-	}
-	return &payouts, total
-}
-
-// GeminiTransformTransactions splits the transactions into appropriately sized blocks for signing
-func GeminiTransformTransactions(ctx context.Context, oauthClientID string, transactions []settlement.Transaction) (*[][]gemini.PayoutPayload, error) {
-	maxCount := 30
-	blocksCount := (len(transactions) / maxCount) + 1
-	privateRequests := make([][]gemini.PayoutPayload, 0)
-	i := 0
-	logger := zerolog.Ctx(ctx)
-
-	txnID := transactions[0].SettlementID
-	txID := uuid.Must(uuid.FromString(txnID))
-	total := decimal.NewFromFloat(0)
-	for i < blocksCount {
-		lowerBound := i * maxCount
-		upperBound := (i + 1) * maxCount
-		payoutLength := len(transactions)
-		if payoutLength <= upperBound {
-			upperBound = payoutLength
-		}
-		transactionBlock := transactions[lowerBound:upperBound]
-		if len(transactionBlock) > 0 {
-			payoutBlock, blockTotal := GeminiConvertTransactionsToGeminiPayouts(&transactionBlock, txnID)
-			total = total.Add(blockTotal)
-			privateRequests = append(privateRequests, *payoutBlock)
-		}
-		i++
-	}
-
-	logger.UpdateContext(func(c zerolog.Context) zerolog.Context {
-		return c.Str("transaction_id", txID.String()).
-			Int("blocks", blocksCount).
-			Int("transactions", len(transactions)).
-			Str("total", total.String())
-	})
-
-	return &privateRequests, nil
 }
